Update default render handling of Merge Records
Previously this was deleting the moving object and updating the
stationary one.  That's almost never going to be how it should be, so
this swaps it.  The stationary object is deleted and the moving object
gets updated.

I think that due to this change, Merge and Split records could be
consolidated into single Records again, but it might make sense to
think about how merging multiple tiles could work before deciding on
that path.  Merges with more than 2 tiles aren't exactly supported by
this approach either, though.